+++ bin/store/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the CRDT Store project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.58
+### Author
+Jean
+
+### Changes
+- ***Timezone***: Added timezone conversion on the query of Find.
+  - Converted from the server timezone to the client's request timezone.
+- ***Date and Time fields***: Fix issue on querying date and time fields on Find.
+---
+
 ## 0.1.57
 ### Author
 Kashan